fix: skpdkb create, and get list fix: get list sptpd and skpdkb query builder
