+++ internal/handlers/bapenda/spt/spt.go
@@ -5,7 +5,6 @@ import (
 	"io"
 	"net/http"
 	"regexp"
-	"strings"
 
 	hj "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/httpjson"
 	rp "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/responses"
@@ -127,6 +126,9 @@ func GetList(w http.ResponseWriter, r *http.Request) {
 		input.Type = mtypes.JenisPajakOA
 	}
 
+	jkOpt := "IS NULL"
+	input.JenisKetetapan_Opt = &jkOpt
+
 	result, err := s.GetList(input, 0, "bapenda", nil)
 	hh.DataResponse(w, result, err)
 }
@@ -238,115 +240,5 @@ func Update(w http.ResponseWriter, r *http.Request) {
 		hj.WriteJSON(w, http.StatusBadRequest, rp.ErrSimple{Message: err.Error()}, nil)
 	}
 	result, err = s.UpdateDetail(id, input, opts)
-	hh.DataResponse(w, result, err)
-}
-
-func Verify(w http.ResponseWriter, r *http.Request) {
-	id, pass := hh.ValidateIdUuid(w, chi.URLParam(r, "id"))
-	if !pass {
-		return
-	}
-
-	var input m.VerifyDto
-	if !hh.ValidateStructByIOR(w, r.Body, &input) {
-		return
-	}
-
-	authInfo := r.Context().Value("authInfo").(*auth.AuthInfo)
-	result, err := s.Verify(id, input, uint(authInfo.User_Id))
-	hh.DataResponse(w, result, err)
-}
-
-func SkpdkbExisting(w http.ResponseWriter, r *http.Request) {
-	typeSpt := strings.ToUpper(chi.URLParam(r, "type"))
-	var input m.SkpdkbExisting
-	err := validateDetail(w, r.Body, &input)
-	if err != nil {
-		return
-	}
-	authInfo := r.Context().Value("authInfo").(*auth.AuthInfo)
-	opts := make(map[string]interface{})
-	opts["userId"] = uint(authInfo.User_Id)
-	opts["newFile"] = false
-	opts["baseUri"] = typeSpt
-	result, err := s.SkpdkbExisting(input, opts)
-	hh.DataResponse(w, result, err)
-}
-
-func SkpdNew(w http.ResponseWriter, r *http.Request) {
-	var input m.Input
-	var err error
-	typeSpt := strings.ToUpper(chi.URLParam(r, "type"))
-	if typeSpt != "SA" {
-		hj.WriteJSON(w, http.StatusBadRequest, rp.ErrSimple{
-			Message: "type tidak valid",
-		}, nil)
-		return
-	}
-	re := regexp.MustCompile(`^\/\w*`)
-	authInfo := r.Context().Value("authInfo").(*auth.AuthInfo)
-	opts := make(map[string]interface{})
-	opts["baseUri"] = re.FindString(r.RequestURI)[1:]
-	opts["userId"] = uint(authInfo.User_Id)
-	opts["newFile"] = true
-	category := r.URL.Query().Get("category")
-	switch category {
-	case "air":
-		var tmp m.CreateDetailAirDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "hiburan":
-		var tmp m.CreateDetailHiburanDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "hotel":
-		var tmp m.CreateDetailHotelDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "parkir":
-		var tmp m.CreateDetailParkirDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "ppjnonpln":
-		var tmp m.CreateDetailPpjNonPlnDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "ppjpln":
-		var tmp m.CreateDetailPpjPlnDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	case "resto":
-		var tmp m.CreateDetailRestoDto
-		err = validateDetail(w, r.Body, &tmp)
-		if err != nil {
-			return
-		}
-		input = &tmp
-	default:
-		err = errors.New("category tidak diketahui")
-		hj.WriteJSON(w, http.StatusBadRequest, rp.ErrSimple{Message: err.Error()}, nil)
-	}
-	if err != nil {
-		return
-	}
-	result, err := s.SkpdkbNew(input, opts)
 	hh.DataResponse(w, result, err)
 }
