+++ internal/services/spt/spt.go
@@ -15,7 +15,6 @@ import (
 
 	srek "github.com/bapenda-kota-malang/apin-backend/internal/services/configuration/rekening"
 	snomertracker "github.com/bapenda-kota-malang/apin-backend/internal/services/spt/sptnomertracker"
-	suser "github.com/bapenda-kota-malang/apin-backend/internal/services/user"
 	a "github.com/bapenda-kota-malang/apin-backend/pkg/apicore"
 	rp "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/responses"
 	t "github.com/bapenda-kota-malang/apin-backend/pkg/apicore/types"
@@ -58,6 +57,10 @@ func filePreProcess(b64String, docsname string, userId uint, oldId uuid.UUID) (f
 	}
 	fileName = sh.GenerateFilename(docsname, id, userId, extFile)
 	return
+}
+
+func generateKodeBilling(kodeBilling string, nomerSpt string) string {
+	return fmt.Sprintf("%s%s", kodeBilling, nomerSpt)
 }
 
 func Create(input m.CreateDto, opts map[string]interface{}, tx *gorm.DB) (any, error) {
@@ -137,7 +140,9 @@ func Create(input m.CreateDto, opts map[string]interface{}, tx *gorm.DB) (any, e
 	yearTwoDigit := yearNow % 1e2
 	nomerSpt := fmt.Sprintf("%d%s", yearTwoDigit, nomerUrut)
 	data.NomorSpt = fmt.Sprintf("%c-%s", kode[0], nomerSpt)
-	data.KodeBilling = fmt.Sprintf("%s%s", *dataRekening.KodeBilling, nomerSpt)
+	if opts["baseUri"].(string) == "sptpd" {
+		data.KodeBilling = generateKodeBilling(*dataRekening.KodeBilling, nomerSpt)
+	}
 	switch data.Type {
 	case mtypes.JenisPajakOA, mtypes.JenisPajakSA:
 	default:
@@ -191,10 +196,18 @@ func GetList(input m.FilterDto, userId uint, cmdBase string, tx *gorm.DB) (any, 
 		endOfMonthDate := datatypes.Date(sh.EndOfMonth(val.(time.Time)))
 		input.JatuhTempo = &endOfMonthDate
 		opts = "="
-	} else {
+	}
+
+	if input.JatuhTempo == nil && input.JenisKetetapan_Opt == nil || *input.JenisKetetapan_Opt == "IS NULL" {
 		dataDateNow := datatypes.Date(time.Now())
 		input.JatuhTempo = &dataDateNow
 	}
+
+	if input.JenisKetetapan_Opt != nil && (*input.JenisKetetapan_Opt == "IS NOT NULL" || *input.JenisKetetapan_Opt == "IS NULL") {
+		baseQuery.Where(fmt.Sprintf("\"JenisKetetapan\" %s", *input.JenisKetetapan_Opt))
+		input.JenisKetetapan_Opt = nil
+	}
+
 	statusAll := ""
 	if input.StatusData != nil {
 		statusData := *input.StatusData
@@ -202,6 +215,7 @@ func GetList(input m.FilterDto, userId uint, cmdBase string, tx *gorm.DB) (any, 
 		if cmdBase == "wp" {
 			switch statusData {
 			case m.TbpStatusFilterBaru, m.TbpStatusFilterPembayaran, m.TbpStatusFilterLunas, m.TbpStatusFilterJatuhTempo:
+				// do nothing
 			default:
 				return sh.SetError("request", "get-data-list", source, "failed", "status data invalid", data)
 			}
@@ -243,6 +257,7 @@ func GetList(input m.FilterDto, userId uint, cmdBase string, tx *gorm.DB) (any, 
 		Scopes(gh.Filter(input)).
 		Count(&count).
 		Scopes(gh.Paginate(input, &pagination)).
+		Order("\"JatuhTempo\" DESC").
 		Find(&data)
 	if result.Error != nil {
 		return sh.SetError("request", "get-data-list", source, "failed", "gagal mengambil data", data)
@@ -416,73 +431,6 @@ func UpdatePenyetoran(id uuid.UUID, tx *gorm.DB) error {
 	return nil
 }
 
-func Verify(id uuid.UUID, input m.VerifyDto, userId uint) (any, error) {
-	var data *m.Spt
-	// validate data exist and copy input (payload) ke struct data jika tidak ada akan error
-	dataRow := a.DB.First(&data, "\"Id\" = ? AND \"Type\" = ?", id.String(), mtypes.JenisPajakOA).RowsAffected
-	if dataRow == 0 {
-		return nil, errors.New("data tidak dapat ditemukan")
-	}
-	if data.StatusPenetapan == mtypes.StatusVerifikasiDisetujuiKabid {
-		return sh.SetError("request", "update-data", source, "failed", "data telah disetujui", data)
-	}
-	if err := sc.Copy(&data, &input); err != nil {
-		return sh.SetError("request", "update-data", source, "failed", "gagal mengambil data payload", data)
-	}
-
-	resp, err := suser.GetJabatanPegawai(userId)
-	if err != nil {
-		return sh.SetError("request", "update-data", source, "failed", "gagal data pegawai: "+err.Error(), data)
-	}
-	jabatan := strings.ToUpper(resp.(string))
-	userRole := ""
-	if kasubid := strings.Contains(jabatan, "KEPALA SUB BIDANG"); kasubid {
-		data.Kasubid_User_Id = &userId
-		userRole = "kasubid"
-	} else if kabid := strings.Contains(jabatan, "KEPALA BIDANG"); kabid {
-		data.Kabid_User_Id = &userId
-		userRole = "kabid"
-	}
-	if userRole == "" {
-		return sh.SetError("request", "update-data", source, "failed", "pegawai bukan kabid atau kasubid", data)
-	}
-	switch input.StatusPenetapan {
-	case "disetujui":
-		if userRole == "kasubid" {
-			data.StatusPenetapan = mtypes.StatusVerifikasiDisetujuiKasubid
-		} else if userRole == "kabid" {
-			data.StatusPenetapan = mtypes.StatusVerifikasiDisetujuiKabid
-		}
-	case "ditolak":
-		if userRole == "kasubid" {
-			data.StatusPenetapan = mtypes.StatusVerifikasiDitolakKasubid
-		} else if userRole == "kabid" {
-			data.StatusPenetapan = mtypes.StatusVerifikasiDitolakKabid
-		}
-	default:
-		return sh.SetError("request", "update-data", source, "failed", "status tidak diketahui", data)
-	}
-	switch data.StatusPenetapan {
-	case mtypes.StatusVerifikasiBaru,
-		mtypes.StatusVerifikasiDisetujuiKasubid,
-		mtypes.StatusVerifikasiDisetujuiKabid,
-		mtypes.StatusVerifikasiDitolakKasubid,
-		mtypes.StatusVerifikasiDitolakKabid:
-		// do nothing
-	default:
-		return sh.SetError("request", "update-data", source, "failed", "status penetapan tidak diketahui", data)
-	}
-	if result := a.DB.Save(&data); result.Error != nil {
-		return sh.SetError("request", "update-data", source, "failed", "gagal mengambil menyimpan data", data)
-	}
-	return rp.OK{
-		Meta: t.IS{
-			"affected": strconv.Itoa(int(dataRow)),
-		},
-		Data: data,
-	}, nil
-}
-
 func Delete(id uuid.UUID) (any, error) {
 	//data spt
 	var data *m.Spt
@@ -504,84 +452,4 @@ func Delete(id uuid.UUID) (any, error) {
 		},
 		Data: data,
 	}, nil
-}
-
-func SkpdkbExisting(input m.SkpdkbExisting, opts map[string]interface{}) (any, error) {
-	var newDataInput m.Input
-	var existingData m.Spt
-	var createdNewData m.Spt
-	err := a.DB.Transaction(func(tx *gorm.DB) error {
-		// get old data from db
-		dataRow := tx.Preload(clause.Associations, func(tx *gorm.DB) *gorm.DB {
-			return tx.Omit("Password")
-		}).First(&existingData, "\"Id\" = ?", input.Spt_Id.String()).RowsAffected
-		if dataRow == 0 {
-			return errors.New("data tidak dapat ditemukan")
-		}
-
-		// casting to dto
-		if existingData.DetailSptAir != nil {
-			newDataInput = &m.CreateDetailAirDto{}
-		} else if existingData.DetailSptHiburan != nil {
-			newDataInput = &m.CreateDetailHiburanDto{}
-		} else if existingData.DetailSptHotel != nil {
-			newDataInput = &m.CreateDetailHotelDto{}
-		} else if existingData.DetailSptParkir != nil {
-			newDataInput = &m.CreateDetailParkirDto{}
-		} else if existingData.DetailSptNonPln != nil {
-			newDataInput = &m.CreateDetailPpjNonPlnDto{}
-		} else if existingData.DetailSptPln != nil {
-			newDataInput = &m.CreateDetailPpjPlnDto{}
-		} else if existingData.DetailSptReklame != nil {
-			newDataInput = &m.CreateDetailReklameDto{}
-		} else if existingData.DetailSptResto != nil {
-			newDataInput = &m.CreateDetailRestoDto{}
-		} else {
-			newDataInput = &m.CreateDetailBaseDto{}
-		}
-		// copy data from existing
-		if err := newDataInput.SkpdkbDuplicate(&existingData, &input); err != nil {
-			return err
-		}
-
-		// calculate skpdkb process
-		newDataInput.CalculateSkpdkb()
-		// create new data
-		respNewData, err := CreateDetail(newDataInput, opts, tx)
-		if err != nil {
-			return err
-		}
-		createdNewData = respNewData.(rp.OKSimple).Data.(m.Spt)
-
-		// change existing data value
-		// existingData.Ref_Spt_Id = &createdNewData.Id
-		// TODO: FLAG SKPDKB
-		// update existing data
-		if result := tx.Save(&existingData); result.Error != nil {
-			return err
-		}
-
-		return nil
-	})
-	if err != nil {
-		return sh.SetError("request", "create-skpdkb-existing", source, "failed", "transaction skpd existing: "+err.Error(), createdNewData)
-	}
-	return rp.OKSimple{Data: createdNewData}, nil
-}
-
-func SkpdkbNew(input m.Input, opts map[string]interface{}) (any, error) {
-	var createdData m.Spt
-	err := a.DB.Transaction(func(tx *gorm.DB) error {
-		// save new data
-		respNewData, err := CreateDetail(input, opts, tx)
-		if err != nil {
-			return err
-		}
-		createdData = respNewData.(rp.OKSimple).Data.(m.Spt)
-		return nil
-	})
-	if err != nil {
-		return sh.SetError("request", "create-skpdkb-new", source, "failed", "transaction skpd new: "+err.Error(), createdData)
-	}
-	return rp.OKSimple{Data: createdData}, nil
 }
